requirements-dev.txt
coverage
flake8 >= 2.0
httmock >= 1.2.1
nose
pylint
scripttest
python-dateutil